newsBrew/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-dotenv==1.0.1
apscheduler==3.10.4
openai==1.51.2
httpx==0.27.2
gspread==6.1.2
google-auth==2.35.0
resend==2.4.0
google-genai==1.73.1
pytest==8.3.3
pytest-asyncio==0.24.0
